Add files via upload
Update on Initial Code Structure. Various comments included throughout to ensure the purpose of code is clear and the relevant additions that need to be made can be seen. The arrows next to different sections of the code indicate significant changes or developments that will take place as the structure of the code develops.