Add files via upload
Code update includes rough code for implementation of BLE turning around procedure and the processing of line break scenarios. Future update will include slope detection implementation. Note: The use of the QEI library for mbed may not have been implemented correctly, so these two procedures above should be reviewed by other members of the team to ensure that the approach is both syntactically correct as well as correct in execution.