Add files via upload
Code now includes the slope detection function which is able to detect inclines or declines in the surface as well as bends in the path the buggy is following. In addition, conditional logic has been implemented to the main function to adjust the speed of the buggy's pwm output depending on the output of the slope detection function. Further adjustments will need to be made depending on the performance of the program in testing and whether it functions as intended.